split data_model_json_fields.go into multiple files
