+++ persistence/immediate_task_info_json.go
@@ -1,0 +1,23 @@
+// Copyright (c) 2023 XDBLab Organization
+// SPDX-License-Identifier: BUSL-1.1
+
+package persistence
+
+import "encoding/json"
+
+type ImmediateTaskInfoJson struct {
+	// used when the `task_type` is waitUntil or execute
+	WorkerTaskBackoffInfo *WorkerTaskBackoffInfoJson `json:"workerTaskBackoffInfo"`
+	// used when the `task_type` is localQueueMessage
+	LocalQueueMessageInfo []LocalQueueMessageInfoJson `json:"localQueueMessageInfo"`
+}
+
+func BytesToImmediateTaskInfo(bytes []byte) (ImmediateTaskInfoJson, error) {
+	var obj ImmediateTaskInfoJson
+	err := json.Unmarshal(bytes, &obj)
+	return obj, err
+}
+
+func FromImmediateTaskInfoIntoBytes(obj ImmediateTaskInfoJson) ([]byte, error) {
+	return json.Marshal(obj)
+}
