+++ persistence/state_execution_sequence_maps_json.go
@@ -1,0 +1,60 @@
+// Copyright (c) 2023 XDBLab Organization
+// SPDX-License-Identifier: BUSL-1.1
+
+package persistence
+
+import (
+	"encoding/json"
+	"fmt"
+)
+
+type StateExecutionSequenceMapsJson struct {
+	SequenceMap map[string]int `json:"sequenceMap"`
+	// store what state execution IDs are currently running
+	// [stateId] -> [stateIdSequence] -> true
+	PendingExecutionMap map[string]map[int]bool `json:"pendingExecutionMap"`
+}
+
+func NewStateExecutionSequenceMapsFromBytes(bytes []byte) (StateExecutionSequenceMapsJson, error) {
+	var seqMaps StateExecutionSequenceMapsJson
+	err := json.Unmarshal(bytes, &seqMaps)
+	return seqMaps, err
+}
+
+func NewStateExecutionSequenceMaps() StateExecutionSequenceMapsJson {
+	return StateExecutionSequenceMapsJson{
+		SequenceMap:         map[string]int{},
+		PendingExecutionMap: map[string]map[int]bool{},
+	}
+}
+
+func (s *StateExecutionSequenceMapsJson) ToBytes() ([]byte, error) {
+	return json.Marshal(s)
+}
+
+func (s *StateExecutionSequenceMapsJson) StartNewStateExecution(stateId string) int {
+	s.SequenceMap[stateId]++
+	seqId := s.SequenceMap[stateId]
+	stateMap, ok := s.PendingExecutionMap[stateId]
+	if ok {
+		stateMap[seqId] = true
+	} else {
+		stateMap = map[int]bool{
+			seqId: true,
+		}
+	}
+	s.PendingExecutionMap[stateId] = stateMap
+	return seqId
+}
+
+func (s *StateExecutionSequenceMapsJson) CompleteNewStateExecution(stateId string, stateSeq int) error {
+	pendingMap, ok := s.PendingExecutionMap[stateId]
+	if !ok || !pendingMap[stateSeq] {
+		return fmt.Errorf("the state is not started, all current running states: %v", pendingMap)
+	}
+	delete(pendingMap, stateSeq)
+	if len(pendingMap) == 0 {
+		delete(s.PendingExecutionMap, stateId)
+	}
+	return nil
+}
