+++ persistence/worker_task_backoff_info_json.go
@@ -1,0 +1,13 @@
+// Copyright (c) 2023 XDBLab Organization
+// SPDX-License-Identifier: BUSL-1.1
+
+package persistence
+
+type WorkerTaskBackoffInfoJson struct {
+	// CompletedAttempts is the number of attempts that have been completed
+	// for calculating next backoff interval
+	CompletedAttempts int32 `json:"completedAttempts"`
+	// FirstAttemptTimestampSeconds is the timestamp of the first attempt
+	// for calculating next backoff interval
+	FirstAttemptTimestampSeconds int64 `json:"firstAttemptTimestampSeconds"`
+}
